kind::unified: Introduce private::SealedContainer and Unified traits.
